+++ .editorconfig
@@ -9,3 +9,6 @@ trim_trailing_whitespace = true
 [{*.json, *.ts, *.tsx, *.css}]
 indent_size = 2
 indent_style = space
+
+[*.md]
+trim_trailing_whitespace = false
