+++ package.json
@@ -35,6 +35,7 @@
     "@types/node": "^24.1.0",
     "@typescript-eslint/eslint-plugin": "^8.0.0",
     "@typescript-eslint/parser": "^8.0.0",
+    "@vitest/coverage-v8": "^3.2.4",
     "domhandler": "^5.0.3",
     "eslint": "^9.0.0",
     "eslint-import-resolver-typescript": "^4.4.4",
